Refactor GPIO clean methods into common code.
