+++ common/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Common code for use across all apps."""
