+++ motor_controller/tests/test_models.py
@@ -9,7 +9,7 @@ from django.core.exceptions import ValidationError
 from django.test import TestCase
 
 # Project
-from constants import AVAILABLE_RPI_GPIO_PINS
+from common.constants import AVAILABLE_RPI_GPIO_PINS
 from motor_controller.constants import STEPPER_DRIVER_TYPES
 from motor_controller.exceptions import CommandError
 from motor_controller.exceptions import ConfigurationError
