+++ switch_controller/constants.py
@@ -2,7 +2,6 @@
 """Constants for switch controller."""
 
 # Project
-import constants as global_constants  # noqa: F401
 
 PUSH_SWITCH_TYPES = [
     ("PTM", "Push To Make"),
